# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhoAmThuc API",
    "description": "Auto-generated OpenAPI spec from Express routes.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/analytics/events": {
      "post": {
        "description": "Upload analytics events in batch.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsEventsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "events"
                ],
                "properties": {
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "deviceId": {
                          "type": "string"
                        },
                        "sessionId": {
                          "type": "string"
                        },
                        "poiId": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string"
                        },
                        "durationMs": {
                          "type": "integer"
                        },
                        "language": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_analytics_events",
        "summary": "Upload analytics events"
      }
    },
    "/api/v1/analytics/presence/heartbeat": {
      "post": {
        "description": "Update online and active user windows by heartbeat.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceHeartbeatResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_analytics_presence_heartbeat",
        "summary": "Send presence heartbeat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "deviceId",
                  "sessionId",
                  "timestamp"
                ],
                "properties": {
                  "deviceId": {
                    "type": "string"
                  },
                  "sessionId": {
                    "type": "string"
                  },
                  "appState": {
                    "type": "string"
                  },
                  "audioState": {
                    "type": "string"
                  },
                  "timestamp": {
                    "type": "integer"
                  },
                  "language": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/stats": {
      "get": {
        "description": "Get analytics summary metrics.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "get_api_v1_analytics_stats",
        "summary": "Get analytics stats"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Validation failed"
              }
            }
          }
        }
      },
...
      "AnalyticsEventsResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "processedCount": {
            "type": "integer"
          },
          "failedCount": {
            "type": "integer"
          }
        }
      },
      "PresenceHeartbeatResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "onlineNowWindowSec": {
            "type": "integer"
          },
          "active5mWindowSec": {
            "type": "integer"
          }
        }
      },
      "AnalyticsStatsResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "plays": {
                "type": "integer"
              },
              "qrScans": {
                "type": "integer"
              },
              "topPois": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "poiId": {
                      "type": "string"
                    },
                    "playCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}